requirements.txt
aiogram>=3.0
python-dotenv
validators
google-api-python-client